Randomize staging IP address to allow parallel usages on same machine
VirtualBox restricts IP addresses of host-only networking adapters to
the range 192.168.56.0/21 (see https://stackoverflow.com/q/69722254).

If there is already an IP address from another run, we must reuse that
because it is engraved into a VM snapshot, if any. This is only needed
if the CLI version is the same, as only then the snapshot is used; so
when reading the Vagrantfile, we only need to handle it in the exact
format written by the current CLI version.