Handle looping commands in pipelines
"The exit status of a pipeline is the exit status of the last command in
the pipeline," with a sensible exception here for invalid programs
(https://www.gnu.org/software/bash/manual/html_node/Pipelines.html).

Also, ensure processes are killed on drop to clean up.

Once could also work with `SIGPIPE` but it is not that simple.